+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Highlights are marked with a pancake 🥞
 
 - Fix missing gossip events in sync manager [#988](https://github.com/p2panda/p2panda/pull/988)
 
+### Changed
+
+- `p2panda-auth` remove high-level API and orderer generic [#1030](https://github.com/p2panda/p2panda/pull/1030)
+
 ## [0.5.1] - 09/02/2026
 
 ### Changed